Split out rdp
